+++ src/main/java/co/nvqa/common/core/cucumber/glue/ApiOrderSteps.java
@@ -306,4 +306,32 @@ public class ApiOrderSteps extends CoreStandardSteps {
     doWithRetry(() -> getOrderClient().updateOrderCod(orderId, newCodAmount),
         "Operator add/update order COD amount:");
   }
+
+  /**
+   * @param dataTableRaw <br/> <b>orderId</b>: {KEY_LIST_OF_CREATED_ORDERS[1].id} <br/>
+   *                     <b>copAmount</b>: 1
+   */
+  @When("API Core - Operator update order COP:")
+  public void apiCoreUpdateOrderCOP(Map<String, String> dataTableRaw) {
+    final Map<String, String> dataTable = resolveKeyValues(dataTableRaw);
+    final Double copAmount = Double.parseDouble(dataTable.get("copAmount"));
+    final Long orderId = Long.parseLong(dataTable.get("orderId"));
+    doWithRetry(() -> getOrderClient().updateOrderCop(orderId, copAmount),
+        "API Core - Operator update order COP");
+  }
+
+  /**
+   * @param dataTableRaw <br><b>trackingId:</b>
+   *                     {KEY_LIST_OF_CREATED_TRACKING_IDS[1]}<br><b>mode:</b> OTP | NONE | AGE |
+   *                     SIGNATURE | IDENTIFICATION<br><i>Choose one of the delivery verification
+   *                     mode</i>
+   */
+  @When("API Core - Update order delivery verification mode:")
+  public void apiCoreUpdateOrderDeliveryVerification(Map<String, String> dataTableRaw) {
+    final Map<String, String> dataTable = resolveKeyValues(dataTableRaw);
+    final String trackingId = dataTable.get("trackingId");
+    final String mode = dataTable.get("mode");
+    doWithRetry(() -> getOrderClient().editDeliveryVerificationRequired(trackingId, mode),
+        "API Core - Update order delivery verification mode");
+  }
 }