+++ src/main/java/co/nvqa/common/core/cucumber/glue/DbCoreSteps.java
@@ -1,7 +1,9 @@
 package co.nvqa.common.core.cucumber.glue;
 
 import co.nvqa.common.core.cucumber.CoreStandardSteps;
+import co.nvqa.common.core.hibernate.CodCollectionDao;
 import co.nvqa.common.core.hibernate.CodInboundsDao;
+import co.nvqa.common.core.hibernate.OrderDao;
 import co.nvqa.common.core.hibernate.OrderDetailsDao;
 import co.nvqa.common.core.hibernate.OrderJaroScoresV2Dao;
 import co.nvqa.common.core.hibernate.OrderTagsDao;
@@ -17,12 +19,14 @@ import co.nvqa.common.core.hibernate.WarehouseSweepsDao;
 import co.nvqa.common.core.hibernate.WaypointsDao;
 import co.nvqa.common.core.model.order.Order;
 import co.nvqa.common.core.model.order.Order.Transaction;
+import co.nvqa.common.core.model.persisted_class.core.CodCollections;
 import co.nvqa.common.core.model.persisted_class.core.CodInbounds;
 import co.nvqa.common.core.model.persisted_class.core.CoreRouteLogs;
 import co.nvqa.common.core.model.persisted_class.core.OrderDetails;
 import co.nvqa.common.core.model.persisted_class.core.OrderJaroScoresV2;
 import co.nvqa.common.core.model.persisted_class.core.OrderTags;
 import co.nvqa.common.core.model.persisted_class.core.OrderTagsSearch;
+import co.nvqa.common.core.model.persisted_class.core.Orders;
 import co.nvqa.common.core.model.persisted_class.core.OutboundScans;
 import co.nvqa.common.core.model.persisted_class.core.Reservations;
 import co.nvqa.common.core.model.persisted_class.core.RouteMonitoringData;
@@ -79,6 +83,11 @@ public class DbCoreSteps extends CoreStandardSteps {
   private OrderTagsSearchDao orderTagsSearchDao;
   @Inject
   private RouteWaypointDao routeWaypointDao;
+  @Inject
+  private CodCollectionDao codCollectionDao;
+
+  @Inject
+  private OrderDao orderDao;
 
   @Override
   public void init() {
@@ -123,8 +132,13 @@ public class DbCoreSteps extends CoreStandardSteps {
       Waypoints result = waypointsDao.getWaypointsDetails(resolvedWayPointIdKey);
       Assertions.assertThat(result.getZoneType())
           .as("Assertion for Zone Type column value is as expected").isEqualTo(expectedZoneType);
-      Assertions.assertThat(result.getRoutingZoneId())
-          .as("Assertion for Zone Id column value is as expected").isNull();
+      if (result.getRoutingZoneId() == null) {
+        Assertions.assertThat(result.getRoutingZoneId())
+            .as("Assertion for Zone Id column value is null as expected").isNull();
+      } else {
+        Assertions.assertThat(result.getRoutingZoneId())
+            .as("Assertion for Zone Id column value is zero as expected").isZero();
+      }
     }, "Validating verified Zone Type value is as expected", 2000, 3);
   }
 
@@ -135,6 +149,28 @@ public class DbCoreSteps extends CoreStandardSteps {
       Waypoints result = waypointsDao.getWaypointsDetails(resolvedWayPointIdKey);
       put(KEY_CORE_WAYPOINT_DETAILS, result);
     }, "get core waypoint details", 2000, 3);
+  }
+
+  @Then("DB Core - verifies that latitude is equal to {string} and longitude is equal to {string} and for waypointId {string}")
+  public void dbCoreVerifiesLatLong(String expectedLatitude, String expectedLongitude,
+      String waypointId) {
+    Long resolvedWayPointIdKey = Long.parseLong(resolveValue(waypointId));
+    doWithRetry(() -> {
+      Waypoints result = waypointsDao.getWaypointsDetails(resolvedWayPointIdKey);
+      String[] formattedValues = formatLatLongValues(result.getLatitude(), result.getLongitude());
+      Assertions.assertThat(formattedValues[0])
+          .as("Assertion for lat column value is as expected").isEqualTo(expectedLatitude);
+      Assertions.assertThat(formattedValues[1])
+          .as("Assertion for lat column value is as expected").isEqualTo(expectedLongitude);
+    }, "Validating verified lat long values are as expected");
+  }
+
+  private String[] formatLatLongValues(Double latitude, Double longitude) {
+    String lat = latitude.toString();
+    String formattedLatitude = lat.substring(0, 6);
+    String lon = longitude.toString();
+    String formattedLongitude = lon.substring(0, 6);
+    return new String[]{formattedLatitude, formattedLongitude};
   }
 
   @When("DB Core - verify route_logs record:")
@@ -530,4 +566,29 @@ public class DbCoreSteps extends CoreStandardSteps {
             })
         , "verify route_waypoint records", 10_000, 3);
   }
+
+  @When("DB Core - get order by order tracking id {string}")
+  public void getOrderDetailByTrackingId(String trackingNumber) {
+    String resolvedTrackingNumber = resolveValue(trackingNumber);
+    doWithRetry(() -> {
+      Orders orders = orderDao.getSingleOrderDetailsByTrackingId(resolvedTrackingNumber);
+      Assertions.assertThat(orders)
+          .withFailMessage("Unexpected order not found with tracking number: %s", trackingNumber)
+          .isNotNull();
+      put(KEY_LIST_OF_CREATED_ORDERS, orders);
+    }, "get order record", 10_000, 3);
+  }
+
+  @And("DB Core - Operator verifies cod_collections record:")
+  public void verifyCodCollections(Map<String, String> data) {
+    CodCollections expected = new CodCollections(resolveKeyValues(data));
+    var actual = codCollectionDao.getMultipleCodCollections(expected.getWaypointId());
+    Assertions.assertThat(actual)
+        .as("List of cod_collections records for waypointId=%s", expected.getWaypointId())
+        .isNotEmpty();
+    Assertions.assertThat(actual)
+        .as("List of cod_collections records for waypointId=%s", expected.getWaypointId())
+        .anyMatch(expected::matchedTo);
+  }
+
 }