+++ CHANGELOG.md
@@ -2,6 +2,46 @@
 
 <hr>
 
+## [1.3.10] 2023-06-15
+
+### Added
+
+- API Core - Operator failed to add reservation to route using data below:
+- API Core - Operator failed to remove reservation id {string} from route
+
+
+## [1.3.9] 2023-06-14
+
+### Added
+
+- DB Core - get order by order tracking id {string}
+
+<hr>
+
+## [1.3.8] 2023-06-08
+
+### Added
+
+- DB Core - verify order id {string} orders.data.belt_dimensions record:
+
+<hr>
+
+## [1.3.7] 2023-06-08
+
+### Added
+
+- API Core - Operator parcel transfer to a new route:
+
+<hr>
+
+## [1.3.6] 2023-06-08
+
+### Added
+
+- API Core - Operator update order COP:
+
+<hr>
+
 ## [1.3.5] 2023-06-05
 
 ### Added